Major update to VGAM library with new class structure

+++ src/fitting/vgam_estimate_curves.R
@@ -3,7 +3,7 @@
 #
 args <- commandArgs(trailingOnly = TRUE)
 if (length(args) == 0) {
-    print("No arguments supplied, use R CMD BATCH \"--args input_file='filename' output_file='filename' degrees_of_freedom=2 save_plot=1 plot_file='filename'\" ref_curve_generator.R")
+    print("No arguments supplied, use R CMD BATCH \"--args input_file='filename' output_file='filename' densities_file='filename' degrees_of_freedom=2 plot_file='filename'\" ref_curve_generator.R")
     quit()
 } else {
     for (i in 1:length(args)) {
@@ -49,37 +49,40 @@ df <- write.csv(fit_frame, file = output_file)
 #
 # Save probability densities to csv file
 #
-min_progression <- min(table[,1])
-max_progression <- max(table[,1])
-progression_step <- 10
-min_value <- min(table[,2])
-max_value <- max(table[,2])
-value_offset <- (max_value - min_value) * 0.2
-value_step <- (max_value - min_value) / 250
-values <- seq(min_value - value_offset, max_value + value_offset, by = value_step)
-values_frame <- data.frame(values)
-for (prog in seq(min_progression, max_progression, by = progression_step)) {
-	deplot(fit, x0 = prog, y = values, show.plot = FALSE) -> aa
-	values_frame <- data.frame(values_frame, aa@post$deplot$density)
+if (exists(densities_file)) {
+	min_progression <- min(table[,1])
+	max_progression <- max(table[,1])
+	progression_step <- 10
+	min_value <- min(table[,2])
+	max_value <- max(table[,2])
+	value_offset <- (max_value - min_value) * 0.2
+	value_step <- (max_value - min_value) / 250
+	values <- seq(min_value - value_offset, max_value + value_offset, by = value_step)
+	values_frame <- data.frame(values)
+	for (prog in seq(min_progression, max_progression, by = progression_step)) {
+		deplot(fit, x0 = prog, y = values, show.plot = FALSE) -> aa
+		values_frame <- data.frame(values_frame, aa@post$deplot$density)
+	}
+	colnames(values_frame) <- c("values",seq(min_progression, max_progression, by = progression_step))
+	df <- write.csv(t(values_frame), file = densities_file)
 }
-colnames(values_frame) <- c("values",seq(min_progression, max_progression, by = progression_step))
-df <- write.csv(t(values_frame), file = densities_file)
-
 
 #
 # Save plot
 #
-w <- 12
-h <- 8
-if (save_plot == 0) {
-    x11(width = w, height = h)
-} else {
-    pdf(plot_file, width = w, height = h)
-}
-qtplot(fit, percentiles = c(5, 25, 50, 75, 95), main = "Quantiles", xlim = c(min_progression, max_progression), las = 1, ylab = "Metric value", lwd = 2, lcol = 4)
-
-if (save_plot == 0) {
-    z <- locator()
-} else {
-    dev.off()
+if (exists(plot_file)) {
+	w <- 12
+	h <- 8
+	if (save_plot == 0) {
+	    x11(width = w, height = h)
+	} else {
+	    pdf(plot_file, width = w, height = h)
+	}
+	qtplot(fit, percentiles = c(5, 25, 50, 75, 95), main = "Quantiles", xlim = c(min_progression, max_progression), las = 1, ylab = "Metric value", lwd = 2, lcol = 4)
+	
+	if (save_plot == 0) {
+	    z <- locator()
+	} else {
+	    dev.off()
+	}
 }
